commits

+++ server/controller/Cerrands.js
@@ -13,12 +13,7 @@ exports.userLogin = async (req, res) => {
       },
     });
     if (result != null) {
-      req.session.user_info = result.dataValues; // v
-      console.log("======================");
-      console.log(req.session.user_info.user_name); // v
-      console.log(req.session.user_info.user_type);
-      console.log("======================");
-      console.log(result.dataValues.user_name);
+      req.session.user_info = result.dataValues;
       console.log(req.session);
       res.send({ user_info: result, msg: true });
     } else {
@@ -37,9 +32,9 @@ exports.checkUserId = async (req, res) => {
     });
     console.log(result);
     if (!result) {
-      return res.send(false);
-    } else {
       return res.send(true);
+    } else {
+      return res.send(false);
     }
   } catch (err) {
     res.send(err);
@@ -50,8 +45,10 @@ exports.checkUserId = async (req, res) => {
 exports.checkUserName = async (req, res) => {
   try {
     const result = await Errands.User_info.findOne({
+      // attributes: ['user_name'],
       where: { user_name: { [Op.eq]: req.body.user_name } },
     });
+    // result
     if (!result) {
       return res.send(true);
     } else {
@@ -68,6 +65,7 @@ exports.userRegister = async (req, res) => {
     const result = await Errands.User_info.findOne({
       where: { user_id: { [Op.eq]: req.body.user_id } },
     });
+    console.log(result);
     if (!result) {
       Errands.User_info.create({
         user_id: req.body.user_id,
@@ -75,8 +73,10 @@ exports.userRegister = async (req, res) => {
         user_name: req.body.user_name,
         user_type: req.body.user_type,
       });
+      res.send(true);
+    } else {
+      res.send(false);
     }
-    res.send(true);
   } catch (err) {
     res.send(err);
   }
@@ -151,34 +151,34 @@ exports.userLike = async (req, res) => {
 // 회원탈퇴
 exports.userWithdrawal = async (req, res) => {
   try {
-    //   const auth = await Errands.User_info.findOne({
-    //      attributes: ['user_name'],
-    //      where: { user_name: { [Op.eq]: req.session.user_name } },
-    //    });
-    //    console.log(req.session.user_name);
-    //    if (auth.dataValues.user_name == req.session.user_name) {
-    const result = await Errands.User_info.destroy({
-      where: { user_id: { [Op.eq]: req.params.userId } },
+    const auth = await Errands.User_info.findOne({
+      attributes: ["user_name"],
+      where: { user_id: { [Op.eq]: req.params.commentId } },
     });
-    console.log(result);
-    if (!result) {
-      res.send(false);
-    } else {
-      req.session.destroy();
-      res.send(true);
+    if (auth.dataValues.user_name == req.session.user_info.user_name) {
+      const result = await Errands.User_info.destroy({
+        where: { user_id: { [Op.eq]: req.params.userId } },
+      });
+      console.log(result);
+      if (!result) {
+        res.send(false);
+      } else {
+        req.session.destroy();
+        res.send(true);
+      }
     }
   } catch (err) {
     res.send(err);
   }
 };
 // 회원정보 수정
-exports.update_user_info = async (req, res) => {
+exports.userUpdate = async (req, res) => {
   try {
     // const auth = await Errands.User_info.findOne({
     //   attributes: ['user_name'],
-    //   where: { user_name: { [Op.eq]: req.session.user_name } },
+    //   where: { user_name: { [Op.eq]: req.session.user_info.user_name } },
     // });
-    // if (auth.dataValues.user_name == req.session.user_name) {
+    // if (auth.dataValues.user_name == req.session.user_info.user_name) {
     const [result] = await Errands.User_info.update(
       {
         user_id: req.body.user_id,
@@ -188,12 +188,37 @@ exports.update_user_info = async (req, res) => {
       },
       { where: { user_id: { [Op.eq]: req.params.userId } } }
     );
-    console.log(result);
-    if (result === 1) {
-      res.send(true);
-    } else {
+    if (result === 0) {
       res.send(false);
+    } else {
+      const update_session = Errands.User_info.findOne({
+        where: {
+          user_id: req.session.user_id,
+          user_pw: req.session.user_pw,
+        },
+      });
+      req.session.user_info = update_session.dataValues;
+      res.send(true);
     }
+  } catch (err) {
+    res.send(err);
+  }
+};
+
+exports.search_wanter_board = async (req, res) => {
+  try {
+    // req.params.search
+    // req.params.optionValue
+    // 작성자 제목 지역
+    // SEQUELIZE LIKE FILTER
+    // 구현해오기
+    // ex)
+    // %검색어% switch
+    // req.params.optionValue = wanter_board_writer
+    const option = req.params.optionValue;
+    const result = await Errands.Wanter_board.findAll({
+      where: { option: { [Op.eq]: req.params.optionValue } },
+    });
   } catch (err) {
     res.send(err);
   }
@@ -328,7 +353,17 @@ exports.delete_wanter_board = async (req, res) => {
 };
 
 // 조회수 up
-// exports.hit_wanter_board = async (req, res) => {};
+exports.hit_wanter_board = async (req, res) => {
+  try {
+    const result = await Errands.Wanter_board.increment(
+      { wanter_board_hit: 1 },
+      { where: { wanter_board_id: { [Op.eq]: req.params.boardId } } }
+    );
+    res.send(result);
+  } catch (err) {
+    res.send(err);
+  }
+};
 
 // ======= Wanter_comment =======
 // 댓글 보여주기
@@ -482,7 +517,7 @@ exports.create_helper_board = async (req, res) => {
       res.send("로그인하시오");
     } else {
       const result = await Errands.Helper_board.create({
-        helper_board_writer: req.session.user_name,
+        helper_board_writer: req.session.user_info.user_name,
         helper_board_title: req.body.helper_board_title,
         helper_board_content: req.body.helper_board_content,
         helper_board_place: req.body.helper_board_place,
@@ -564,7 +599,17 @@ exports.delete_helper_board = async (req, res) => {
 };
 
 // 조회수 up
-// exports.hit_helper_board = async (req, res) => {};
+exports.hit_helper_board = async (req, res) => {
+  try {
+    const result = await Errands.Helper_board.increment(
+      { helper_board_hit: 1 },
+      { where: { helper_board_id: { [Op.eq]: req.params.boardId } } }
+    );
+    res.send(result);
+  } catch (err) {
+    res.send(err);
+  }
+};
 
 // ======= Helper_comment =======
 // 댓글 보여주기
@@ -711,20 +756,24 @@ exports.read_one_notice = async (req, res) => {
 // 공지 생성
 exports.create_notice = async (req, res) => {
   try {
-    const auth = await Errands.User_info.findOne({
-      attributes: ["user_type"],
-      where: { user_type: { [Op.eq]: req.session.user_type } },
-    });
-    if (auth.dataValues.user_type == "root") {
-      const result = Errands.Notice.create({
-        notice_writer: req.session.user_name,
-        notice_title: req.body.notice_title,
-        notice_content: req.body.notice_content,
-      });
-
-      res.send(result);
+    if (!req.session.user_info) {
+      res.send("로그인하시오");
     } else {
-      res.send("생성 권한이 없습니다");
+      const auth = await Errands.User_info.findOne({
+        attributes: ["user_type"],
+        where: { user_type: { [Op.eq]: req.session.user_info.user_type } },
+      });
+      if (auth.dataValues.user_type == "root") {
+        const result = Errands.Notice.create({
+          notice_writer: req.session.user_name,
+          notice_title: req.body.notice_title,
+          notice_content: req.body.notice_content,
+        });
+
+        res.send(result);
+      } else {
+        res.send("생성 권한이 없습니다");
+      }
     }
   } catch (err) {
     res.send(err);
@@ -734,24 +783,28 @@ exports.create_notice = async (req, res) => {
 // 공지 수정
 exports.update_notice = async (req, res) => {
   try {
-    const auth = await Errands.User_info.findOne({
-      attributes: ["user_type"],
-      where: { user_type: { [Op.eq]: req.session.user_type } },
-    });
-    if (auth.dataValues.user_type == "root") {
-      const [result] = Errands.Notice.update(
-        {
-          notice_title: req.body.notice_title,
-          notice_content: req.body.notice_content,
-        },
-        { where: { notice_id: { [Op.eq]: req.params.boardId } } }
-      );
-      if (result === 0) {
-        return res.send(false);
-      }
-      res.send(true);
+    if (!req.session.user_info) {
+      res.send("로그인하시오");
     } else {
-      res.send("수정 권한이 없습니다");
+      const auth = await Errands.User_info.findOne({
+        attributes: ["user_type"],
+        where: { user_type: { [Op.eq]: req.session.user_info.user_type } },
+      });
+      if (auth.dataValues.user_type == "root") {
+        const [result] = Errands.Notice.update(
+          {
+            notice_title: req.body.notice_title,
+            notice_content: req.body.notice_content,
+          },
+          { where: { notice_id: { [Op.eq]: req.params.boardId } } }
+        );
+        if (result === 0) {
+          return res.send(false);
+        }
+        res.send(true);
+      } else {
+        res.send("수정 권한이 없습니다");
+      }
     }
   } catch (err) {
     res.send(err);
@@ -761,23 +814,27 @@ exports.update_notice = async (req, res) => {
 // 공지 삭제
 exports.delete_notice = async (req, res) => {
   try {
-    const auth = await Errands.User_info.findOne({
-      attributes: ["user_type"],
-      where: { user_type: { [Op.eq]: req.session.user_type } },
-    });
-    console.log(auth);
-    if (auth) {
-      if (auth.dataValues.user_type == "root") {
-        const result = Errands.Notice.destroy({
-          where: { notice_id: { [Op.eq]: req.params.boardId } },
-        });
-        if (!result) {
-          return res.send(false);
-        }
-        return res.send(true);
-      }
+    if (!req.session.user_info) {
+      res.send("로그인하시요");
     } else {
-      res.send("삭제 권한이 없습니다");
+      const auth = await Errands.User_info.findOne({
+        attributes: ["user_type"],
+        where: { user_type: { [Op.eq]: req.session.user_info.user_type } },
+      });
+      console.log(auth);
+      if (auth) {
+        if (auth.dataValues.user_type == "root") {
+          const result = Errands.Notice.destroy({
+            where: { notice_id: { [Op.eq]: req.params.boardId } },
+          });
+          if (!result) {
+            return res.send(false);
+          }
+          return res.send(true);
+        }
+      } else {
+        res.send("삭제 권한이 없습니다");
+      }
     }
   } catch (err) {
     res.send(err);
